pyproject.toml
[project]
name = "poll-agent"
version = "0.1.0"
description = "Automated poll topic generator powered by Google Agent SDK (ADK) and Grok"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.20.0",
    "xai-sdk>=1.4.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.0",
    "litellm>=1.80.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/poll_agent"]

[tool.uv]
dev-dependencies = []

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py310"
